Move sudoloop code to exec.go
